+++ lucene/suggest/src/java/org/apache/lucene/search/suggest/tst/TernaryTreeNode.java
@@ -57,7 +57,7 @@ public class TernaryTreeNode {
       mem +=
           RamUsageEstimator.shallowSizeOf(token)
               + RamUsageEstimator.NUM_BYTES_ARRAY_HEADER
-              + Character.BYTES * (long) token.length();
+              + Character.BYTES * token.length();
     }
     mem += RamUsageEstimator.shallowSizeOf(val);
     return mem;
